admin: move the /state JSON visitors into their own header
The two visitors that serialize /state were anonymous-namespace classes inside
StateHandler.cpp, so nothing could construct one -- a test that wanted to check
the JSON had to go through the HTTP handler.

Code motion only: they move to admin/JsonStateVisitors.h verbatim, and
StateHandler includes it. /state output is unchanged.
